fix : customerdata랑 characterdata 하나로 합침
