+++ Assets/General/Scripts/DataClasses/CharacterData.cs
@@ -1,7 +1,26 @@
 using UnityEngine;
 using System.Collections.Generic;
 
-[CreateAssetMenu(fileName = "NewCharacterData", menuName = "Affinity/Character Data")]
+//customerdata에서 가져온 요소들
+// 스프라이트 세트로 저장
+[System.Serializable]
+public class CharacterPose
+{
+    [Tooltip("포즈(감정)에 붙힐 이름")]
+    public string poseName;
+
+    [Tooltip("몸통 부분에 해당하는 스프라이트")]
+    public Sprite bodySprite;
+
+    [Tooltip("뜬 눈 스프라이트")]
+    public Sprite eyesOpenSprite;
+
+    [Tooltip("눈을 깜빡일 때 사용할 감은 눈 스프라이트")]
+    public Sprite eyesClosedSprite;
+}
+
+
+[CreateAssetMenu(fileName = "NewCharacterData", menuName = "Game/Character Data")]
 public class CharacterData : ScriptableObject
 {
     [Header("기본 정보")]
@@ -21,5 +40,11 @@ public class CharacterData : ScriptableObject
 
     [Header("변화하는 값")]
     public bool hasMet;                // 만난 여부 (게임 첫 실행 시에만 쓰임.)
-    [Range(0, 100)] public int affinity; // 호감도 (0.5 단위 = 0~20), (변화하는 요소, 게임 첫 실행 시에만 쓰임.)
+    [Range(0, 100)] public int affinity; // 호감도 (0.5 단위 = 0~20), (변화하는 요소, 게임 첫 실행 시에만 쓰임.)\
+
+
+    //customerdata에서 가져온 요소
+    [Header("포즈 세트(감정 별)")]
+    [Tooltip("이 캐릭터가 가질 수 있는 모든 감정 표현의 리스트")]
+    public List<CharacterPose> poses;
 }
